requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
Markdown==3.7
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
whitenoise==6.8.2
